Refactor per @sarayourfriend's suggestions
